# Optimize contiguous shard and select