--- a/src/datasets/info.py
+++ b/src/datasets/info.py
@@ -211,10 +211,10 @@ class DatasetInfo:
     @classmethod
     def from_merge(cls, dataset_infos: List["DatasetInfo"]):
         dataset_infos = [dset_info.copy() for dset_info in dataset_infos if dset_info is not None]
-        description = "\n\n".join(unique_values(info.description for info in dataset_infos))
-        citation = "\n\n".join(unique_values(info.citation for info in dataset_infos))
-        homepage = "\n\n".join(unique_values(info.homepage for info in dataset_infos))
-        license = "\n\n".join(unique_values(info.license for info in dataset_infos))
+        description = "\n\n".join(unique_values(info.description for info in dataset_infos)).strip()
+        citation = "\n\n".join(unique_values(info.citation for info in dataset_infos)).strip()
+        homepage = "\n\n".join(unique_values(info.homepage for info in dataset_infos)).strip()
+        license = "\n\n".join(unique_values(info.license for info in dataset_infos)).strip()
         features = None
         supervised_keys = None
         task_templates = None