@lhoestq 2022-06-14
I adressed your comments @albertvillanova , let me know what you think :)